+++ app/page.tsx
@@ -21,7 +21,10 @@ export default async function Home() {
               className="mb-2 transition-transform duration-300 will-change-transform transform hover:scale-[1.02] hover:-translate-y-1 shadow-md"
               key={node.path}
             >
-              <Link className="group" href={`/posts/${node.path}` }>
+              <Link
+                className="group"
+                href={`/posts/${node.path}`}
+              >
                 <h2 className="group-hover:text-green-700 transition duration-300 border-b-2 border-gray-300 pb-1 mb-1 text-lg ">
                   {node.name}
                 </h2>