+++ next.config.js
@@ -2,9 +2,9 @@
 const nextConfig = {
   reactStrictMode: true,
   pageExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  experimental: {
-    appDir: true
-  }
+  // experimental: {
+  //   appDir: true
+  // }
 }
 
 module.exports = nextConfig